Added a linker flag so the systemC library can be found in alternative places.
The user can configure the path of essential libraries (systemC, 3d-ice and
superLu) using environment variables. The linker flag added tells the linker
where the dynamic linked library systemC (.so) is to be found. This is not
necessary for static linked libraries (.a) like 3d-ice and superLu.

Readme file updated.